Fix decimal headcount (#594)
Reads headcount as double from DB

+++ src/main/java/com/mercadolibre/planning/model/api/client/db/repository/forecast/ProcessingDistributionView.java
@@ -16,7 +16,7 @@ public interface ProcessingDistributionView {
 
   ProcessName getProcessName();
 
-  long getQuantity();
+  double getQuantity();
 
   MetricUnit getQuantityMetricUnit();
