+++ src/test/java/com/mercadolibre/planning/model/api/usecase/ProcessingDistributionViewImpl.java
@@ -24,7 +24,7 @@ public class ProcessingDistributionViewImpl implements ProcessingDistributionVie
 
     private final ProcessName processName;
 
-    private final long quantity;
+    private final double quantity;
 
     private final MetricUnit quantityMetricUnit;
